:sparkles: api: move session repository to src/certification/session directory
