+++ api/scripts/certification/send-certification-result-emails.js
@@ -4,7 +4,7 @@ dotenv.config();
 
 import bluebird from 'bluebird';
 import { logger } from '../../lib/infrastructure/logger.js';
-import * as sessionRepository from '../../lib/infrastructure/repositories/sessions/session-repository.js';
+import * as sessionRepository from '../../src/certification/session/infrastructure/repositories/session-repository.js';
 import * as certificationCenterRepository from '../../lib/infrastructure/repositories/certification-center-repository.js';
 import * as mailService from '../../lib/domain/services/mail-service.js';
 import { manageEmails } from '../../lib/domain/services/session-publication-service.js';
