+++ api/tests/certification/session/integration/infrastructure/repositories/session-repository_test.js
@@ -1,8 +1,8 @@
-import { databaseBuilder, expect, knex, domainBuilder, catchErr } from '../../../../test-helper.js';
+import { databaseBuilder, expect, knex, domainBuilder, catchErr } from '../../../../../test-helper.js';
 import _ from 'lodash';
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import { Session, statuses } from '../../../../../lib/domain/models/Session.js';
-import * as sessionRepository from '../../../../../lib/infrastructure/repositories/sessions/session-repository.js';
+import { NotFoundError } from '../../../../../../lib/domain/errors.js';
+import { Session, statuses } from '../../../../../../lib/domain/models/Session.js';
+import * as sessionRepository from '../../../../../../src/certification/session/infrastructure/repositories/session-repository.js';
 
 describe('Integration | Repository | Session', function () {
   describe('#save', function () {
