+++ api/tests/integration/domain/usecases/get-attendance-sheet/get-attendance-sheet_test.js
@@ -9,7 +9,7 @@ import { expect, databaseBuilder } from '../../../../test-helper.js';
 import * as writeOdsUtils from '../../../../../lib/infrastructure/utils/ods/write-ods-utils.js';
 import * as sessionXmlService from '../../../../../lib/domain/services/session-xml-service.js';
 import * as readOdsUtils from '../../../../../lib/infrastructure/utils/ods/read-ods-utils.js';
-import * as sessionRepository from '../../../../../lib/infrastructure/repositories/sessions/session-repository.js';
+import * as sessionRepository from '../../../../../src/certification/session/infrastructure/repositories/session-repository.js';
 import * as sessionForAttendanceSheetRepository from '../../../../../lib/infrastructure/repositories/sessions/session-for-attendance-sheet-repository.js';
 import { getAttendanceSheet } from '../../../../../lib/domain/usecases/get-attendance-sheet.js';
 import * as url from 'url';
